requirements.txt
numpy>=1.23
pyyaml>=6.0
omegaconf>=2.3
tqdm>=4.66
scikit-learn>=1.3
torch>=2.1
